# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.City:
    properties:
      alias:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      province:
        $ref: '#/definitions/models.Province'
      province_id:
        type: integer
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.Country:
    properties:
      alias:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      provinces:
        items:
          $ref: '#/definitions/models.Province'
        type: array
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  models.Province:
    properties:
      alias:
        type: string
      cities:
        items:
          $ref: '#/definitions/models.City'
        type: array
      country:
        $ref: '#/definitions/models.Country'
      country_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
info:
  contact: {}
  description: Swagger documentation for reservation API .
  license:
    name: MIT
  title: Reservation API
  version: "1.0"
paths:
  /cities:
    get:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.City'
            type: array
      summary: findAll
      tags:
      - City
    post:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: City
        in: body
        name: city
        required: true
        schema:
          $ref: '#/definitions/models.City'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.City'
      summary: create
      tags:
      - city
  /cities/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: Id
        in: path
        name: Id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.City'
      summary: findById
      tags:
      - City
    put:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: Id
        in: path
        name: Id
        required: true
        type: integer
      - description: City
        in: body
        name: City
        required: true
        schema:
          $ref: '#/definitions/models.City'
      - description: City
        in: body
        name: country
        required: true
        schema:
          $ref: '#/definitions/models.City'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.City'
      summary: update
      tags:
      - City
  /countries:
    get:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Country'
            type: array
      summary: findAll
      tags:
      - Country
    post:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: Country
        in: body
        name: country
        required: true
        schema:
          $ref: '#/definitions/models.Country'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Country'
      summary: create
      tags:
      - Country
  /countries/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: Id
        in: path
        name: Id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Country'
      summary: findById
      tags:
      - Country
    put:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: Id
        in: path
        name: Id
        required: true
        type: integer
      - description: Country
        in: body
        name: Country
        required: true
        schema:
          $ref: '#/definitions/models.Country'
      - description: Country
        in: body
        name: country
        required: true
        schema:
          $ref: '#/definitions/models.Country'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Country'
      summary: update
      tags:
      - Country
  /countries/{id}/provinces:
    get:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: Id
        in: path
        name: Id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Province'
            type: array
      summary: find Provinces by country ID
      tags:
      - Country
...
swagger: "2.0"